rename wiki.py to nytimes.py
